+++ back/src/server/routes/users.js
@@ -50,27 +50,6 @@ router.get('/contacts/new', async (req, res) => {
   }
 })
 
-router.post('/contacts', async (req, res) => {
-  const { fullName, login, number, isAuth } = req.body;
-  try {
-    const userId = jwt.decode(isAuth)._id;
-    const currentUser = await User.findOne({ _id: userId });
-    const newContact = await User.findOne({ login });
-    currentUser.friends.push({
-      fullName,
-      friendId: newContact._id
-    })
-    await currentUser.save();
-    const chat = new Chat({
-      members: [userId, newContact._id]
-    })
-    await chat.save()
-    res.status(201).json({ chatId: chat._id });
-  } catch (error) {
-    res.status(404).send(error);
-  }
-})
-
 router.post('/contacts/all', async (req, res) => {
   const { isAuth } = req.body;
   try {
